Added new NearbyCard
