+++ src/app/components/Amenities/Amenities.module.css
@@ -49,3 +49,65 @@
   }
 }
 
+.flexWrapper {
+  margin: 0; /* Equivalent to m-0 */
+  padding: 0; /* Equivalent to p-0 */
+  display: flex; /* Equivalent to flex */
+  flex-wrap: wrap; /* Equivalent to flex-wrap */
+}
+
+/* List Styles */
+.amenitiesList {
+  margin-top: 30px;
+  display: flex;
+  flex-wrap: wrap;
+  gap: 10px; /* Add space between list items */
+}
+
+/* Individual Item Style */
+.amenityItem {
+  list-style-type: none;
+  width: 50%; /* Default width for small screens */
+  padding: 18px 0 ;
+  font-size: 14px;
+  color: #808080;
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+}
+
+@media (min-width: 640px) {
+  .amenityItem {
+    width: 33.33%; /* Medium-sized screens */
+  }
+}
+
+@media (min-width: 768px) {
+  .amenityItem {
+    width: 25%; /* Larger screens */
+  }
+}
+
+@media (min-width: 1024px) {
+  .amenityItem {
+    width: 20%; /* Extra-large screens */
+  }
+}
+
+/* Icon Styles */
+.icon {
+    font-size: 16px;
+  color: #4caf50; /* Green color for icon */
+  margin-right: 8px;
+  
+}
+
+@media (min-width: 768px) {
+  .icon {
+    width: 20px; /* Larger icon size for medium screens */
+  }
+}
+
+.span{
+  font-size: 15px;
+}
