+++ src/app/components/Amenities/Amenities.tsx
@@ -1,7 +1,9 @@
 "use client";
 import React, { useEffect, useState } from "react";
 import styles from "./Amenities.module.css";
-import AmenitiesCard from "./AmenitiesCard";
+import { IoMdCheckmark } from "react-icons/io";
+
+type AmenitiesType = string[];
 
 const Amenities: React.FC = () => {
   return (
@@ -19,3 +21,35 @@ const Amenities: React.FC = () => {
 };
 
 export default Amenities;
+
+const AmenitiesCard: React.FC = () => {
+  const [Amenities, setAmenities] = useState<AmenitiesType>([]);
+
+  useEffect(() => {
+    const Gallery = async () => {
+      try {
+        const response = await fetch("/api/AmenitiesData");
+        if (!response.ok) {
+          throw new Error("Failed to fetch feature data");
+        }
+        const data: AmenitiesType = await response.json();
+        setAmenities(data);
+      } catch (error) {
+        console.error("Error fetching feature data:", error);
+      }
+    };
+    Gallery();
+  }, []);
+  return (
+    <>
+      <ul className={styles.flexWrapper}>
+        {Amenities.map((item) => (
+          <li className={styles.amenityItem} key={item}>
+            <IoMdCheckmark className={styles.icon} />
+            <span className={styles.span}>{item}</span>
+          </li>
+        ))}
+      </ul>
+    </>
+  );
+};
